+++ src/ConsoleCmdQuantumElevators.cs
@@ -15,7 +15,7 @@ namespace QuantumElevators
         {
             var dict = new Dictionary<string, string>() {
                 { "", "enable/disable debug logging for this mod" },
-                { "push <x> <y> <z>", "recursively push entities 1 block at a time to make room" },
+                { "push <x> <y> <z>", "recursively push entities 1 block at a time to make room; use float values from 'lp' command to let system calculate correct block coordinates for you" },
             };
 
             var i = 1; var j = 1;
@@ -29,7 +29,7 @@ namespace QuantumElevators
 
         public override string GetDescription()
         {
-            return "Enable/Disable debug logging for Quantum Elevators";
+            return "Admin commands for working with Quantum Elevators";
         }
 
         public override string GetHelp()
@@ -53,15 +53,21 @@ namespace QuantumElevators
                         SdtdConsole.Instance.Output("Wrong number of arguments, expected 4, found " + _params.Count.ToString() + ".");
                         return;
                     }
-                    if (!int.TryParse(_params[1], out var x)
-                        || !int.TryParse(_params[2], out var y)
-                        || !int.TryParse(_params[3], out var z))
+                    if (!float.TryParse(_params[1], out var x)
+                        || !float.TryParse(_params[2], out var y)
+                        || !float.TryParse(_params[3], out var z))
                     {
-                        SdtdConsole.Instance.Output("Wrong type of argument, provided string could not be converted to int.");
+                        SdtdConsole.Instance.Output($"Wrong type of argument, provided string could not be converted to float: {_params[1]}, {_params[2]}, {_params[3]}.");
                         return;
                     }
                     SdtdConsole.Instance.Output($"Pushing...");
-                    _ = CoreLogic.Push(new Vector3i(x, y, z)); // TODO: enable or remove command
+                    // note: one can use integer conversion to get the bottom-left of a positive coordinate,
+                    // but not for the bottom-left of a negative coordinate... these have to be rounded "down" to the *next* nearest number.
+                    // ... yes, this is how block coords are identified lol
+                    _ = CoreLogic.Push(new Vector3i(
+                        x % 1 < 0 ? x - 1 : x,
+                        y % 1 < 0 ? y - 1 : y,
+                        z % 1 < 0 ? z - 1 : z));
                     SdtdConsole.Instance.Output($"Pushing completed.");
                     return;
             }